Move config schema generation API out of compiler plugin
