+++ misc/ballerina-config-schema-builder/build.gradle
@@ -1,0 +1,20 @@
+apply from: "$rootDir/gradle/javaProject.gradle"
+apply from: "$rootDir/gradle/repositories.gradle"
+
+dependencies {
+    implementation project(':ballerina-lang')
+    implementation 'com.google.code.gson:gson:2.8.7'
+}
+
+description = 'Ballerina - JSON schema builder for configurable variables'
+
+ext.moduleName = 'ballerina-config-schema-builder'
+
+compileJava {
+    doFirst {
+        options.compilerArgs = [
+                '--module-path', classpath.asPath,
+        ]
+        classpath = files()
+    }
+}
