+++ misc/ballerina-config-schema-builder/src/main/java/module-info.java
@@ -1,0 +1,5 @@
+module org.ballerinalang.config.schema.builder {
+    requires io.ballerina.lang;
+    requires com.google.gson;
+    exports org.ballerinalang.config.schema.builder;
+}
