+++ misc/compiler-plugins/modules/configurable-schema-generator/src/main/java/module-info.java
@@ -1,7 +1,8 @@
-module org.ballerinalang.pkg.toml.semntic.analyzer {
+module org.ballerinalang.config.schema.generator {
     requires io.ballerina.lang;
     requires io.ballerina.tools.api;
     requires io.ballerina.toml;
     requires com.google.gson;
-    exports org.ballerinalang.configschemagenerator;
+    requires org.ballerinalang.config.schema.builder;
+    exports org.ballerinalang.config.schema.generator;
 }
