+++ misc/compiler-plugins/modules/configurable-schema-generator/src/main/java/org/ballerinalang/config/schema/generator/ConfigSchemaGenerator.java
@@ -15,7 +15,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.ballerinalang.configschemagenerator;
+package org.ballerinalang.config.schema.generator;
 
 import io.ballerina.projects.plugins.CodeAnalysisContext;
 import io.ballerina.projects.plugins.CodeAnalyzer;
