+++ misc/compiler-plugins/modules/configurable-schema-generator/src/main/resources/META-INF/services/io.ballerina.projects.plugins.CompilerPlugin
@@ -1,1 +1,1 @@
-org.ballerinalang.configschemagenerator.ConfigSchemaGenerator
+org.ballerinalang.config.schema.generator.ConfigSchemaGenerator
