+++ settings.gradle
@@ -61,6 +61,7 @@ include(':docerina-gradle-plugin')
 include(':docerina')
 include(':ballerina-formatter')
 include(':ballerinalang-data-mapper')
+include ':ballerina-config-schema-builder'
 include(':testerina:testerina-core')
 include(':testerina')
 include(':testerina:report-tools')
@@ -188,6 +189,7 @@ project(':ballerina-shell:shell-rt').projectDir = file('ballerina-shell/modules/
 project(':ballerina-shell:shell-core').projectDir = file('ballerina-shell/modules/shell-core')
 project(':ballerina-shell:shell-cli').projectDir = file('ballerina-shell/modules/shell-cli')
 project(':ballerinalang-data-mapper').projectDir = file('misc/ballerinalang-data-mapper')
+project(':ballerina-config-schema-builder').projectDir = file('misc/ballerina-config-schema-builder')
 project(':docerina-gradle-plugin').projectDir = file('misc/ballerina-gradle-plugins/docerina-gradle-plugin')
 project(':docerina').projectDir = file('misc/docerina')
 project(':testerina:testerina-core').projectDir = file('misc/testerina/modules/testerina-core')
